refactor: move webdriver specs (#165)
* refactor: move wdio tests to __wdio__

We were putting Jest and WDIO tests in the same folder. By separating
them, we can add clarity to the developer and we can remove some ignore directives.

* refactor(jest): update jest configuration

Remove `testPathIgnorePatterns` and instead use `__tests__` folder for
all tests.

* refactor(wdio): update wdio.conf.js

Use the `__wdio__` folder instead of `__tests__`.

* test(integration): remove unused snapshot

This snapshot is no longer used.

* test(integration): fix matcher registration

I don't know why, but the `jest-setup` would not register a matcher.

This was my workaround. I'll test again before we do a release.